fix "done" btn & add new component
